Cosmetic changes to MySharesOrdering
For Yes/No markets: I cloned the title, and added it to the header line of the table for the market, so that the name is in a consistent location.

For all markets: I moved up/down/collapse to the header instead of the footer. Changed the "collapse" feature to pretty symbols instead of words. Enabled collapse for yes/no markets as well as for linked markets. This is in anticipation of collapsing the "margin-title" class upon collapse as well. 

Small modifications within the makeUp, makeDown, makeCollapse functions to prevent repeated calls to getName

To be done in next iteration: upon collapse, hide the cells in the "contract-header" other than title, buttons, clone "risk in investment" element and put that in its place.